agents/osint/osint-agent/requirements.txt
agno>=1.1.1
irc3>=1.1.10
asyncio
aiofiles
pydantic>=2.0.0
python-dotenv
requests
dnspython
openai>=1.0.0